+++ R/piecepackr-package.R
@@ -1,0 +1,12 @@
+#' @section Package options:
+#' The following piecepackr function arguments may be set globally via [base::options()]:
+#'  \describe{
+#'    \item{piecepackr.cfg}{Sets a new default for the `cfg` argument}
+#'    \item{piecepackr.default.units}{Sets a new default for the `default.units` argument}
+#'    \item{piecepackr.envir}{Sets a new default for the `envir` argument}
+#'    \item{piecepackr.op_angle}{Sets a new default for the `op_angle` argument}
+#'    \item{piecepackr.op_scale}{Sets a new default for the `op_scale` argument}
+#'    \item{piecepackr.trans}{Sets a new default for the `trans` argument}
+#'  }
+#' @keywords internal
+"_PACKAGE"